move initial files from private personal repository
Moving the necessary and current files from the private, personal repository with initial development progress. A lot of unnecessary learning hurdles in that original commit history and all future development will be under this public repository. There are a couple files that are in need of updates, but not used by the package at the moment.