refactor & improve plotting methods
Instead of putting all plot code in one script, I'm creating a separate module that can be imported and used with all the various plots to increase modularity and hopefully make it easier to contribute new plots. These also contain some updates for the generic plots. Removed style sheet & add a function to set these parameters. Renamed generate_plots.py to plotter.py & updated to use new plotlib for 5p plot. Others to be re-added, but this commit was starting to get quite large in terms of number of changes..