SummaryStats has been corrected to write one 5' nucleotide length matrix per library. Previously these stats were being merged and summed across libraries.
Strangely, there seems to be a bug in cwltool which results in the URL encoding of special characters in output filenames. This is likely because internally they use file URI's (the "location" field of the CWL File class) and they miss the step of renaming the encoded filename back to its basename after relocating outputs.