The DESeq step is now reintegrated with the pipeline. Counter outputs now go into their own subdirectory.
Added --relax-path-checks to the CWL runners' arguments because Counter outputs containing special characters were being flagged for containing parentheses. This is contrary to the documentation for the feature. "Relaxing" path checks actually throws the door wide open and permits all characters in paths (regex .*). This should not be an issue in Unix systems, which is fine because cwltool no longer supports Windows.