Bugfix for statistics.py which was not correctly defaulting to ID Attribute if an alias didn't exist for a feature. This issue only arose in cases where the user ONLY defined "ID" for their Name Attributes.
Reverting relative import changes as these behave differently when a package component is directly invoked as a script (yikes, what the heck Python...) The dependencies of counter do not need to be reverted because they are not directly invoked by command line.

Also made small changes for the "ID Attribute" -> "Name Attribute" rename, and updated docstrings accordingly.