Resuming pipeline at the Counter step is now fully functional for both native and subprocess CWL runners.
DGE outputs are placed in their own folder. Both the Counter and DGE outputs produced from a "resume" run are timestamped. The timestamp of the `run_name`, which is applied to the contents in these directories, is now updated to reflect the "resume" timestamp.

A copy of the features sheet is included in the new Counter directory because users will most likely use this feature while tweaking their feature selection rules. This means that we should provide a copy of the features sheet used for every "resume" run