Workflow and CommandLineTool CWL has been updated for aquatx-plot. The `counts` and `counts-prep` steps in aquatx_wf.cwl have been renamed to `counter` and `counter-prep` for naming consistency.
Configuration.py has been updated to include the plotter step in ResumeConfig. Comments in the ResumeConfig class have been updated accordingly. It has also been lightly refactored to make its procedures less specific to the counter step and more generic. This is in preparation for allowing the pipeline to be resumed at the plotter step.

The subdirs step has been modified to include features_csv in the counter output directory be default, for all run types. This is no longer exclusive to `aquatx resume`

plotter.py:
- Minor bugfixes and formatting changes

plotterlib.py:
- Figures/axes are now reused between plots for class_pie_barh() and scatter_simple(). Axes are simply cleared between runs. This has had a positive impact on runtime and memory performance.