Bugfixes.
aquatx-deseq.r: the --vanilla option has been removed from the script shebang. This is for compatibility and is temporary until we find a better solution. There is inconsistency in /usr/bin/env among Unix systems. Relatively new Linux distros will fail on this line.

resume.py: `aquatx replot` was crashing if expected pipeline outputs weren't present. When globbing for these outputs, an empty list was returned and indexed at 0; we now catch this IndexError and treat it the same as a FileNotFound exception.

setup.py: the matplotlib stylesheet was not being correctly included in the package data. We now include the entire extras directory.