Configuration updated to handle control condition.
Also added a convenience wrapper class for CSV reading which cleans up the code and makes CSV field specification consistent across the project. This was initiated because originally I planned to pass the Samples Sheet to Plotter to specify control. Changed plans in favor of simply specifying the control condition name as a string on the command line. Keeping the wrapper class because it will be useful for future development, particularly the validation class.