In the spirit of keeping dependencies consistent and under one roof, the following dependencies are now handled by conda rather than pip:  - cwltool  - psutil  - htseq
Nodejs and bowtie are now more strictly version locked to reduce conda's problem size. This will help speed up the installation a bit but at the end of the day it's still going to be a little slower with the dependency additions listed above.

Additionally, the commented toil dependencies have been removed. We have achieved workflow parallelization in a variety of other ways.