During a previous iteration of Plotter, a warning about an unset locale environment variable was produced specifically when running via CWL. I addressed this by simply setting the locale environment variable to US english.
I removed this and instead considered passing the --preserve-entire-environment option to the CWL runner so that locale and all other variables are preserved.

However, on second thought this could produce some weird results in Plotter because matplotlib utilizes environment variables which might change its behavior. If users need to pass options to matplotlib they should do so via the styles sheet.