bowtie-build and fastp have been converted to write their outputs to logfiles. Since bowtie-build is not library specific, its console output file is simply "console_output.log". fastp prepends this with the basename of its input file.
I have also taken the opportunity to clean up our Configuration/cwl routine. Inputs that are essentially "input file basename with stuff appended" now have that performed by the corresponding Workflow. This helps clean up the code for Configuration and helps reduce those huge blocks of input variables in the CWL workflows. This leaves a responsibility for somebody to generate the "basename with stuff" filenames, and I'm having the Workflow classes do that rather than the CommandLineTools. This leaves the CommandLineTools with a non-pipeline-associated user experience which is the whole point.