The routine for determining appropriate DESeq2/Bioconductor version has been improved.
If the setup script detects that the user has DESeq2 installed on the host, it will ask the user if they want to use the already installed version to save time. If yes, neither R nor DESeq2 will be installed in the tinyRNA environment.

The environment creation step now writes Conda output to both the console and the logfile so that users have a visual indication of progress.

Bugfix: shell determination is now more reliable on Mac. If user has changed their default shell via `chsh` then $SHELL does not update accordingly.

Bugfix: detection of an installed Conda client works for bash + Anaconda combo where conda initialization is actually written to .bash_profile instead of .bashrc