The 'Unassigned' category has been added to the subplots in 'class_charts'. This category will only be shown if there are unassigned counts.
The table in 'class_charts' is now properly appended to the pie chart rather than as its own subplot. It is located below the pie chart and now adjusts to width of the widest cell. This removed the need to call .axis('off') each time the table subplot was cleared.

The pie chart legend has been moved to the upper right