Plotter functions have been updated to handle tagged feature counts. I tried to do this with backward compatibility in mind so that users can still run tiny-replot on old outputs. Initial runs look good, but I haven't tested the backward compatibility yet.
Internally raw_counts and norm_counts are multiindex dataframes if they have a Tag column.

Classes are now sorted prior to plotting in sample_avg_scatter_by_dge_class. This was done in order to be consistent with the behavior of class_charts, because it looks so much nicer to have tagged classes adjacent to one another, and because it ensures that tagged classes will be colored differently from each other if the list is long and the colorwheel wraps around.