Finalizing conda lock files. environment.yml has also been updated to include an environment variable which should help with installation stability on Linux. Lock files will not convey environment variables so setup.sh must do so.
If a user had previously installed any component in the dependency tree via pip --user, these components end up in ~/.local/lib/python3.x/site-packages and Python actually gives this directory higher priority in sys.path than the environment's site-packages. Importing a package that can be found in the .local directory will result in that package being preferred even if it is the wrong version. Wild. This strikes me as a bug in conda because it breaks conda's isolation model. Others have thought so as well, but each time this subject has been raised it is ultimately dismissed as intended functionality.