Native Python efficiency improvements:
After doing a lot of profiling and fiddling with the routine in FeatureSelector.choose(), I'm once again back to flattening the list of Stage 2 candidates via a list comprehension. 1/2 of Stage 2 is performed in the comp., and step 2/2 (calculating min_rank) is now performed with one call on the resulting list. While runtimes are better with this approach, elimination by hierarchy is still flawed.

I've also removed the inner generator expression in FeatureCounter.assign_features(). After the recent switch to calling union(*) on StepVector steps, it really doesn't make sense to screen for zero length step sets. This is much more efficient.

Strands are now represented with a boolean value which simplifies some things. Strand selection is now performed via a chain of bitwise XOR operations. This prevents us from having to form and pass a tuple for each evaluation.