FeatureSelector.choose() has been updated to consider hierarchy in Stage 3 selection by sorting Stage 2 matches by their hierarchy value. Once a feature-rule pair passes Stage 3 selection, it sets min_rank to which all subsequent matches are compared. The Stage 3 loop exits as soon as a feature-rule pair is found to differ from min_rank.
Strand representation in tiny-count has been converted to boolean values: True for '+' and False for '-'. This can later be extended to include None for '.', but this currently isn't needed.

The strand selector has been similarly updated. Strand matches are now determined by a chain of boolean XOR operations which are substantially more efficient.