Updated the URL in lockfiles for bioconductor-genomeinfodbdata to use the latest build, which fixes an issue in the recipe that prevented tinyRNA installation (many thanks to Tai for tracking this down).
I've also added a script to automate the process of updating the lockfiles per the contents of environment.yml, which will also include updated build numbers for the entire dependency tree. The script can be used to remedy future issues like this. Note that it was NOT used for the lockfile changes included in this commit since the build updates were quite extensive and I have time limitations for testing this week