Reliability improvements for the installation and use of build dependencies on macOS.
setup.py:
Now treats the Cython StepVector and tests as optional extensions so that installation doesn't error out when there are unforeseen system incompatibilities. Installation also determines the path to the macOS SDK in a way that will be compatible with conda-build pipelines (if/when we have a bioconda recipe for tinyRNA). Also: better condition checks for determining when we need to search for the SDK, and better comments/docstrings.

setup.sh:
CommandLineTools is now checked for and installed if not present because it is required for the C++ compiler on macOS. This marks another curious exception to the Conda isolation model. As far as I can tell, build dependencies on Linux are already covered by Conda subdependencies.