Drafted integration of the new GFFValidation class into pipeline startup (configuration.py) and tiny-count startup (counter.py)
GFF validation is treated as an optional step that must be specifically requested in configuration.py. This is because we will assume that resume runs are using inputs that have already been validated.

GFF validation is skipped in tiny-count during pipeline runs. This is because we will assume that both end-to-end runs and resume runs are using inputs that have already been validated.