Adding execute_post_run_tasks() to Configuration. This function calls others that verify bowtie-build outputs were produced, updates index paths if long indexes were produced, then saves updates to the Paths Sheet and Run Config.
This addresses a long running problem where end-to-end runs with a bowtie-build step would not save the updated ebwt in Paths Sheet if a downstream step produced an error. Now, the updated ebwt path is written to the Paths File at the end of any run where at least bowtie-build ran successfully.