class_charts and scatter_by_dge_class have been updated to treat the Tag multiindex column as a classifier. Wow, this opened the door for some really satisfying simplifications to the class-related codebase.
Backward compatibility is not offered with this commit. I've been thinking about how to reconcile this but I think ultimately that will be a bad idea. Number one, the counting semantics of the old Feature Class vs Tag columns are completely different; these inputs wouldn't be interchangeable.

I've also removed the show_unknown option from scatter_dges() because we have yet to use this feature and we've never added the option to any user-facing config files