Updates to ConfigBase.joinpath() to make it more robust against inputs that are Nonetype.
Updates to ConfigBase.from_here() to allow it to handle inputs that are already CWL file classes, and inputs that are Nonetype or empty string. Previously, this function would return the config file's directory if destination="" or None. In this context this isn't very helpful. Now it will return an empty string.