ConfigBase.from_here() has slightly cleaner logic and is much more resilient.
A CWL file object entry is now included for "paths_file" in the Run Config. This means there will be two keys containing the path to the Paths File; one of them is at the top and is a regular string for user friendliness, and the other is in the automatic section for the workflow runner.

Replaced a missing "_" from the run_directory name that had been recently removed.