Bugfix to restore sanity to ticks and tick labels in scatter plots. Mainly, this entails: 1. A flexible n in "every nth label" that scales with the axis range rather than being a constant value 2. The function that performs the initial calculation of major tick locations no longer places ticks outside of the plot's view limits as it had under certain conditions
Additionally, the lowermost tick mark/label is no longer hidden by default, and the uppermost tick mark/label is no longer hidden if it is too close to the end of the axis. These measures were originally implemented when we were using exponent tick labels which are much wider, and aren't necessary with the shorter log2 labels.