Adding --dge-min and --dge-max optional arguments to tiny-plot. Users are free to specify either or both values. Values are expected to match what the user sees on the labeled plot axes, i.e. the log2 normalized value rather than the raw count value.
This commit also addresses a bug in the lower limit calculation that occurs when the feature with the lowest count also has a count of zero in another sample/group (for DGE scatter plots and replicate scatter plots, respectively). Under certain conditions this could lead to the lowest count feature being clipped from the plot space. This is usually a non-issue due to the axes margins.

get_scatter_view_lims() will report a warning if the values for axes.xmargin and axes.ymargin differ. The larger of the two values will be chosen if this happens.