Updating the Features Sheet to use "nested" instead of "full" (missed this one earlier)
Also adding a (very basic) backward compatibility check for the "full" overlap selector. This should be temporary. A backward compatibility class should be added for the Features Sheet, but in order to do it right, I'd also need to write a Features Sheet class where validation takes place. Need to maintain a consistent design pattern. I need to make a clear delineation between validation and backward compatibility, and that needs to be a separate GH issue.