Cleaning up troublesome uses of dict.get(key) and dict.get(key, default).
If the KEY exists then:
- .get(key) will return its value, otherwise None
- .get(key, default) will return its value, otherwise default

What these methods DON'T do is change based on the VALUE. They don't return None/default if the value is None, '', or otherwise evaluates to False in a boolean context. It seems I've beem assuming the opposite. Embarrassing.

Fortunately all of these corrections are for very unlikely cases (things like users deleting expected keys or changing their type, or features with a Parent= tag that is missing a value). The most likely case is when setting the default value for Run Config's `run_name`, which is pretty harmless.