The _parse_alignments() method of SAM_reader has been replaced by a dedicated Cython extension class. The extension acts as an iterable over pysam alignments that have been converted to dictionary form on the fly. Importantly, it allows us to eliminate nearly all calls to the Python-space API for pysam and instead retrieve information directly from Cython space.
It is much faster. The runtime cost of switching to pysam is now a mere 5% slowdown (down from 20-30% initially).

In the spirit of minimizing Cython footprint due to debugging complications, the extension class still accumulates alignments for decollapsing, but uses a callback method from SAM_reader's Python space to actually write the decollapsed alignments