Lots of updates to setup.sh.
Most importantly, in-place updates to the tinyRNA environment are no longer supported. The environment must be removed and recreated. When this functionality was originally added, dependency updates were taking place within the same Python version. Now they are across Python versions, and this can cause many, many problems. I tried several approaches to preserve this functionality, and while some got close, they were not stable or reliable. This is the safest way.

Additionally:
- Updated the version of miniconda that the script installs (if necessary)
- The Apple Silicon version of Miniconda is installed, if appropriate and necessary
- The Miniconda installer is executed with bash as it appears the latest version doesn't work with zsh
- If Miniconda was installed, the user is instructed to source their shell's .rc file to finalize installation
- It is now possible to run the script from outside of the project directory